+++ server/src/api/models/interfaces/IShortUrl.ts
@@ -1,5 +1,9 @@
-export interface IShortUrl {
+import { Document } from "mongoose";
+
+interface IShortUrl extends Document {
   short: string;
   long: string;
   createdAt: Date;
 }
+
+export default IShortUrl;